wip: some work on JSON History
